refactor: split db tests to multiple files
