+++ packages/database/tests/postgres.test.ts
@@ -1,7 +1,10 @@
 /* eslint-disable jest/no-hooks, jest/valid-describe */
 import testSchema from './test-schema'
 import { DB, PostgresConnector } from '~database'
-import DBTests from './database'
+import FindTests from './database/find'
+import CreateTests from './database/create'
+import UpdateTests from './database/update'
+import DeleteTests from './database/delete'
 
 describe('postgres tests', function(this: any) {
   this.db = {} as DB
@@ -21,5 +24,8 @@ describe('postgres tests', function(this: any) {
     await this.db.close()
   })
 
-  DBTests.bind(this)()
+  FindTests.bind(this)()
+  CreateTests.bind(this)()
+  UpdateTests.bind(this)()
+  DeleteTests.bind(this)()
 })
